backend/requirements/common.txt
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
